+++ src/main/nest/global/decorator/index.ts
@@ -1,4 +1,5 @@
 export * from './api-create-post.decorator';
 export * from './api-update-post.decorator';
 export * from './api-guard.decorator';
+export * from './api-guard-optional.decorator';
 export * from './get-user-id.decorator';
